change default color options
Now the colors will appear only if the command runs in a
terminal and the output is not piped to another command
(like less).

Ofcourse, if the client is running in windows, the colors
will still be disabled, as windows do not understand ANSI
escape codes as unix systems do, which is required for this
implementation of colored output.