Add an environment variable for project slug to be used in tests.
Project slug is unique, so configuration that work with CI settings will
not work locally with the developer's account. So we need to overwrite
it in order integration tests to pass locally.

With this change, the TRANSIFEX_PROJECT environment variable should be
set in the environment, and it's no longer hard coded in the test
script.

Additionally, we fail the build if the necessary environment variables
aren't set and we are building the mainline and correct a bug in the
windows test build script.

Co-Authored-By: Dionysis Tsoumas <diotsoumas@transifex.com>
Co-Authored-By: Konstantinos Koukopoulos <kouk@transifex.com>